refactor existing search bar tests
